pyproject.toml
[project]
name = "egon-content"
version = "0.1.0"
description = "CLI tool to generate mental health Markdown articles for Logseq or Obsidian"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "openai>=1.70.0",
    "typer>=0.15.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
addopts = "--cov=. --cov-report=term-missing --cov-report=xml"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    "tests/*",
    ".venv/*",
]
